upcoming: [DI-22132] - Use common typography
